refactor(getVersion): rename and extract parser function
